Create Group analysis_simple functions
In the table above, movie budgets are represented in USD. Write the function print_budget_in_euros() that will convert those dollar amounts into euros and print the results on the screen. Use the following conversion rate: $1 = €0.90. In the precode, we've defined this first (euros_per_dollar); even though the variable isn't declared within the function, the function can still use it.
The function should accept the dollar amounts as an argument and print results in the following format:
Budget: 181.00 mil. € 
Note: The result is printed precise to the second decimal place.