Create Function return value
1. Let’s try to calculate the total sums of the budgets and box office gross of the films in the table. Write the function column_sum() that accepts the table and column number and returns the sum value of that column. Use this function to print the result in the following format:
Total budget: $... mil.
Total box office gross: $... mil. 
Round the totals to the second decimal place. As a reminder: the budgets are saved in the column indexed 5 and the box office gross is in the column indexed 6.

2. Let’s imagine that every single person watched every Oscar-winning film on the silver screen. In this case, how much would every ticket to every film have to cost, for the films to make at least as much as their current gross?
In order to evaluate those ticket costs, we will need the sum of all the films’ grosses (column indexed 6) divided by the world’s total population. One source the US Census Bureau states that the world’s population in 2018 was at 7,539 million people. You can calculate the total box office gross using the function from the previous task. Divide one by the other and print the results on the screen in the following format:
Every person spent an average of $... on Oscar-winning films. 
Just as you did before, round the result to two decimal places.

3. The function print_mean() that you wrote in the previous lesson calculates an average value and prints it on the screen. Change this function so that instead of printing the average it returns the average.
In order to calculate the sum, you won't need to form a loop; call the function column_sum() instead. Recycling code is good for the planet.