Create Functions as a code-organizing tool
Task:

Now let's calculate the per-minute cost for each film. In order to do this, write the function add_cost_per_minute(): it should add the value it calculates as a new column.
The function should accept a single argument: the data table. This is the table that the function will then change. We have already written the code for printing output on the screen, complete it by including the following equation:
COST_PER_MINUTE = BUDGET / LENGTH
(Remember, the length is in minutes already.)