Create Conditional statement
Task:

Write the function check_if_long() that will determine whether or not a film is lengthy. We'll consider any film less than two hours long as short and everything else as long.
The function should accept the length of a film in minutes as input and then print either
Long film.
or
Short film.
on the screen.