Create Table filtering
Task:

1. Filter the table, keeping films from the years 2000 and later, then sort it in descending order by box office gross and print the top 5 results.

2. Write a function that filters out films longer than a certain value.
The function should have two input arguments: the table and the specified running time. It should return a new table, instead of changing the original.
In the new table, you must only include films that are two hours or less and sort them by box office gross in descending order. Then print the top 5 elements of the returned table on the screen.