Create The "in" operator
1. Let's filter the table by genre. To do so, write the function filter_by_genre(). It should include two arguments: data (you'll specify oscar_data when you call the function) and genre (the genre you want to filter by). It should return a completely new table that includes the films from the genre that you specify. Remember that the function is universal, and the arguments you use when defining it are abstract. When you call the function, you enter particular arguments relevant to your data: the name of the table you want this function to process and the genre you want to filter by.
Once you have defined the function, call it with "melodrama" as the second argument.

2. Some genres are more common than others. Let's try to find the most popular ones.
You'll create an empty table, saving it in the variable genres_counts . Then you will fill it with data using a loop. It will have columns containing genre names and corresponding numbers of films.
We've collected all the genres in the variable all_genres. Loop through every genre, using our filter_by_genre function from the previous task to find the number of films in each genre and adding this information to the table. To do so, call the filter_by_genre function in the parentheses of len(). Save the result in the variable count.
First fill in the table with data, then sort the number column in descending order. We've already written code to display data in the following format:
Genre        | Number
------------------------
...          |        ...