Create Multiple functions within a program
Task:

Expand the data by adding ROI and Cost per minute columns. To do so call add_roi() and then add_cost_per_minute().
Paying close attention to the precode, call column_mean() inside the for cycle to calculate the mean values for rating, running time, ROI, and cost per minute for each genre of interest. Notice how the column_sum() is nested inside the column_mean() function.
(Here, I had a bit of pre-code). The precode will compile the information into this table:
Genre      | Rating | Length  | ROI  | Cost per minute
-------------------------------------------------------
history    |    ... |    ...  |  ... |              ...
... 
Save it in the genres_means variable and print it on the screen. Print all the numbers rounded to the second decimal place and align them to the right while aligning the text to the left.