Create The "not" operator
Task:

Finish the function filter_without_genre(). It should toss all the films of a specified genre from the table to the cutting room floor. The arguments of this function are: the table of films and the genre that needs to be excluded. Don't change the source data; create a new table within the function and use it to return the result.
Call up the written function to find all non-drama films.