Create Abbreviating Conditions
Task:

Write the function filter_by_year() to select films according to year. It should return a table of films released within a specified year range.
The function should have three inputs: the dataset, the first year of the desired period, and the year after the end of the desired period. (If you set 1980-1985, films released over a five-year period are included, starting with 1980 and ending with 1984, excluding 1985. This is similar to Python indexing, which includes the first value of a range but not the end value.) Use abbreviated coding for your conditions.
We have already written the code for printing the table.